+++ src/app/components/main-components/news-center/news-center.component.ts
@@ -7,7 +7,7 @@ import {Post, UsersAndPosts} from '../../../models/model classes/posts/Post';
 import {User} from '../../../models/model classes/user/User';
 import {catchError, map, mergeMap, takeUntil} from 'rxjs/operators';
 import {Title} from '@angular/platform-browser';
-import {PaginatorComponent} from '../../../shared/components/paginator/paginator.component';
+import {PaginatorComponent} from '../../../shared/ui-components/paginator/paginator.component';
 import {of, Subject} from 'rxjs';
 import {NgxSpinnerService} from 'ngx-spinner';
 import {ModalPopupService} from "../../../shared/services/modal-popup.service";
