+++ src/app/shared/ui-components/address-auto-completer/address-auto-completer.component.scss
