+++ src/app/shared/ui-components/horizontal-scrollable-tabs/horizontal-scrollable-tabs.component.spec.ts
