+++ src/app/shared/ui-components/paginator/paginator.component.scss
