+++ src/app/shared/ui-components/row-text-editor/row-text-editor.component.scss
