+++ src/app/shared/ui-components/segmented-control/segmented-control.component.scss
