+++ src/app/shared/ui-components/segmented-control/segmented-control.component.spec.ts
