+++ src/app/shared/ui-components/text-pair-editor/text-pair-editor.component.scss
